% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_states_geofacet_grid_4x4.R
\name{de_states_geofacet_grid_4x4}
\alias{de_states_geofacet_grid_4x4}
\title{Geographical ordering of states for faceting using geofacet}
\usage{
de_states_geofacet_grid_4x4(linebreak = FALSE)
}
\arguments{
\item{linebreak}{Should longer names be broken up into two lines?}
}
\description{
This function returns a dataframe of state IDs, state names and locations on a 4x4 grid to be used for faceting plot panels using geofacet.
}
