% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_stateparl.R
\docType{data}
\name{link_stateparl}
\alias{link_stateparl}
\title{ID Links between bundeslaendeR data and parliamentary speech data from
StateParl dataset}
\format{
A tibble containing one row per party group in StateParl data.
\describe{
\item{state}{chr ISO 3166-2:DE-code of the state in StateParl data}
\item{period}{dbl Electoral Period in StateParldataset}
\item{affiliation}{chr Party affiliation of speaker in StateParl dataset.}
\item{state_bundeslaender}{chr State name abbreviation in bundeslaendeR data.}
\item{election_date}{dbl Election date in bundeslaendeR data}
\item{partyname_short}{chr Party name in bundeslaendeR data.}
\item{link_notes}{chr Notes on link. Primarily concerned with party groups that are founded later on in the electoral cycle and therefore cannot be matched to election data.}
}
}
\usage{
data(link_stateparl)
}
\description{
Dataset providing a link between ltw_elections (or ltw_combined) and data on
parliamentary speeches from the StateParl database (paragraphs dataset). Party groups that are
founded later in the legislative cycle cannot be linked to electoral outcome
data. See link_notes for details on such cases.
}
\details{
Note that StateParl treats the time before and after the re-election of the Berlin
Abgeordnetenhaus in 2023 (re-running the 2021 election) as the same electoral period.
In the bundeslaendeR data the state_election_term count gets incremented.
The election_date variable in link_stateparl thus contains both election dates for this
electoral period, separated by a semi-colon. Also use the the date variable from StateParl's paragraphs dataset to link the data for this electoral period.

Note that StateParl's affiliation variable denotes the political position
instead of the party affiliation for some speeches, in particular for
members of the state parliament presidency and -more crucially- for members
of the state government. link_stateparl thus cannot not help with linking these politicians'
speeches to party level election results, etc.

Beltermann, E., Souris, A., Nguyen, C., & Kropp, S. (2025). StateParl (Version
2.0.0) Dataset. GESIS, Köln. https://doi.org/10.7802/2854
}
\examples{
\dontrun{


link_stateparl \%>\%
 separate(
   col = election_date,
   into = c("election_date_1", "election_date_2"),
   sep = ";"
 ) \%>\%
 pivot_longer(
   cols = c(election_date_1, election_date_2),
   values_to = "election_date",
   names_to = "tmp"
 ) \%>\%
 filter(!is.na(election_date)) \%>\%
 select(-tmp) \%>\%
 mutate(election_date = as.Date(election_date)) \%>\%
 left_join(
   ltw_elections,
   by = c("state_bundeslaender" = "state", "election_date", "partyname_short")
 )
}

}
\keyword{datasets}
