% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltw_combined.R
\docType{data}
\name{ltw_combined}
\alias{ltw_combined}
\title{Election Results of Landtagswahlen in the German states (Bundeslaender) since 1946 as well as government constellation.}
\format{
A tibble containing one row per contesting party per election and per cabinet.
\describe{
\item{state}{chr ISO 3166-2:DE-code of the state; including BA for the former state of Baden, WH for the former state of W端rttemberg-Hohenzollern and WB for the former state of W端rttemberg-Baden}
\item{nuts1}{chr NUTS1 code of state. NA for former states Baden, W端rttemberg-Baden, W端rttemberg-Hohenzollern.}
\item{state_name_de}{chr German name of the state}
\item{state_name_en}{chr English name of the state}
\item{state_election_term}{dbl Election term in the state}
\item{election_date}{date Election date}
\item{election_id_bundeswahlleiter}{chr Specific election_id as denoted by the Bundeswahlleiter. Note that BA, WH and WH are named as BW and the number counts down. NA for cases taken from Landeswahlleiters (ie elections after ST 2021).}
\item{election_remarks_bundeswahlleiter}{chr Remarks on the election as given by the Bundeswahlleiter}
\item{electorate}{dbl Number of eligible voters. For more totals also see the last three columns.}
\item{number_of_voters}{dbl Number of voters turning out. For more totals also see the last three columns.}
\item{turnout}{dbl Turnout. Share of eligible voters turning out.}
\item{valid_votes}{dbl Number of valid votes. Must not be equal to the number of ballots cast, as sometimes a ballot contains multiple votes! For more totals also see the last three columns.}
\item{total_seats_parliament}{dbl Total number of members of the newly elected Landtag.}
\item{female_party_seats_available }{lgl Denotes whether information on the no. of female members of the Landtag per party is available. Note that for parties not elected to the new Landtag party_female_mps always is.na() == TRUE.}
\item{total_female_mps_parliament}{dbl Number of newly elected female MPs.}
\item{partyname_short}{chr Harmonized abbreviation of the party's name.}
\item{partyname}{chr Harmonized name of the party.}
\item{partyname_short_bundeswahlleiter}{chr Partyname abbreviation as documented by the Bundeswahlleiter.}
\item{partyname_bundeswahlleiter}{chr Partyname as documented by the Bundeswahlleiter.}
\item{party_vote_count}{dbl Number of votes recieved by the party.}
\item{party_vshare}{dbl Share of votes recieved by the party.}
\item{party_seat_count}{dbl Number of seats recieved by the party.}
\item{party_sshare}{dbl Share of seats recieved by the party.}
\item{party_female_mps}{dbl Number of female MPs elected for the party. Note that for parties not elected to the new Landtag party_female_mps always is.na() == TRUE.}
\item{wzb_govelec_id}{dbl If available, MR-Code of the party in the internal govelec database of the WZB department Democracy and Democratization}
\item{ches_id}{dbl If available, ID of the party in the Chapel-Hill Expert Survey.}
\item{partyfacts_id}{dbl If available, ID of the party in the partyfacts database.}
\item{decker_neu}{lgl Denotes, wether the Handbuch der deutschen Parteien (3. ed.) by Decker and Neu has a chapter on the party.}
\item{url_info}{chr URL to informaton on the party on the web. Can contain multiple URLs!}
\item{party_remarks_stelzle}{chr Remarks on the party by me.}
\item{party_remarks_bundeswahlleiter}{chr Remarks on the party as listed by the Bundeswahlleiter.}
\item{gueltige_stimmzettel_hh_hb}{chr Messy totals.}
\item{gesamtstimmen_by}{chr Messy totals.}
\item{ausgefallene_stimmen_be}{chr Messy totals.}
\item{abgegebene_stimmen_hh}{chr Messy totals.}
\item{ungueltige_stimmen_except_hh_hb}{chr Messy totals.}
\item{ungueltige_stimmzettel_hh_hb}{chr Messy totals.}
\item{gov_no_within_legterm}{int Order of cabinets within the legislative session of the state.}
\item{gov_id}{dbl Unique ID of government. Taken from Linhart et al. However, this ID is not counting up within state by time. In cases where Governments were missing from Linhart et al. before the timeframe covered by Linhart et al. (eg. in Berlin) these earlier governments have an higher ID than later cabinets contained in Linhart et al. data.}
\item{state_gov_number}{int Order of cabinets within the state.}
\item{gov_start_date}{int Start date of the government.}
\item{gov_party}{lgl Boolean wether the party was a cabinet party. Note: There is a single cabinet where no party is marked as part of the cabinet: Heinrich Welsch's caretaker government in the Saarland (at the time not yet a member of the FRG) in 1955.}
\item{nmin_party}{dbl Number of ministers of the party in the cabinet. NA for parties not in government!}
\item{gov_source}{chr Source of the information on the government. Either Linhart et al. or the URL of the German Wikipedia Page containing information on the cabinet.}
\item{gov_remarks_stelzle}{chr My remarks on governments.}
\item{minister_president}{chr Name of minister president}
\item{mp_party}{chr Party of the minister president. partyname_short format used. Note: There is a single cabinet with an independent minister president: Heinrich Welsch's caretaker government in the Saarland (at the time not yet a member of the FRG) in 1955.}
\item{is_mp_party}{lgl Is the governments minister president from this party? Note: There is a single cabinet where the minister president is not part of any party: Heinrich Welsch's caretaker government in the Saarland (at the time not yet a member of the FRG) in 1955.}
}
}
\source{
Election results data provided by the Bundeswahlleiter. A machine-readable version of the data in the pdf available here (\url{https://www.bundeswahlleiter.de/service/landtagswahlen.html}) was kindly provided to me. Election data outside the timeframe covered by Bundeswahlleiter's data provided to me was collected from the states' local election authorities' (Landeswahlleiter) websites. More information on parties and the continuity of parties under different labels was collected by me. Information on Governments mainly taken from the replication data of Linhart, Eric, Franz U. Pappi und Ralf Schmitt (2008): Die proportionale Ministerienaufteilung in deutschen Koalitionsregierungen: Akzeptierte Norm oder das Ausnutzen strategischer Vorteile?, Politische Vierteljahresschrift 49(1): 46-67. To be found online here: \url{https://www.tu-chemnitz.de/phil/politik/pspi/forschung/daten.php}. Information outside the timeframe of Linhart et al. was collected by me, mainly from German Wikipedia.
}
\usage{
data(ltw_combined)
}
\description{
Dataset containing election results of Landtagswahlen in the German states (Bundeslaender) since 1946 as well as linked information on government constellation.
}
\keyword{datasets}
